old implementation on my gpu:
```
step = 1/10, RSS = 3498.251 MB, Δt = 11.940 s
step = 2/10, RSS = 3498.251 MB, Δt = 11.887 s
step = 3/10, RSS = 3498.251 MB, Δt = 12.155 s
```
new implementation on my gpu:
```
step = 1/10, RSS = 3229.816 MB, Δt = 0.904 s
step = 2/10, RSS = 3229.816 MB, Δt = 0.727 s
step = 3/10, RSS = 3229.816 MB, Δt = 0.698 s
step = 4/10, RSS = 3229.816 MB, Δt = 0.698 s
step = 5/10, RSS = 3229.816 MB, Δt = 0.698 s
step = 6/10, RSS = 3229.816 MB, Δt = 0.697 s
step = 7/10, RSS = 3229.816 MB, Δt = 0.697 s
step = 8/10, RSS = 3229.816 MB, Δt = 0.697 s
step = 9/10, RSS = 3229.816 MB, Δt = 0.698 s
step = 10/10, RSS = 3229.816 MB, Δt = 0.699 s
```

old implementation on cluster gpu:
```
step = 1/10, RSS = 3498.251 MB, Δt = 0.940 s
step = 2/10, RSS = 3498.251 MB, Δt = 0.827 s
step = 3/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 4/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 5/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 6/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 7/10, RSS = 3498.251 MB, Δt = 0.829 s
step = 8/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 9/10, RSS = 3498.251 MB, Δt = 0.828 s
step = 10/10, RSS = 3498.251 MB, Δt = 0.828 s
```
new implementation on cluster gpu:
```
step = 1/10, RSS = 3229.816 MB, Δt = 0.898 s
step = 2/10, RSS = 3229.816 MB, Δt = 0.475 s
step = 3/10, RSS = 3229.816 MB, Δt = 0.468 s
step = 4/10, RSS = 3229.816 MB, Δt = 0.467 s
step = 5/10, RSS = 3229.816 MB, Δt = 0.465 s
step = 6/10, RSS = 3229.816 MB, Δt = 0.461 s
step = 7/10, RSS = 3229.816 MB, Δt = 0.461 s
step = 8/10, RSS = 3229.816 MB, Δt = 0.461 s
step = 9/10, RSS = 3229.816 MB, Δt = 0.461 s
step = 10/10, RSS = 3229.816 MB, Δt = 0.461 s
```

In both cases, the new version is faster. What's astounding is how bad the matmul implementation is my home gpu.